Refactored to use Stencil generator
