+++ src/components/form-controls/int-select-item/int-select-item.css
@@ -1,0 +1,3 @@
+:host {
+  display: block;
+}
