+++ src/components/form-controls/int-select-item/int-select-item.spec.ts
@@ -1,0 +1,7 @@
+import { IntSelectItem } from './int-select-item';
+
+describe('int-select-item', () => {
+  it('builds', () => {
+    expect(new IntSelectItem()).toBeTruthy();
+  });
+});
