+++ src/components/form-controls/int-select/int-select.css
@@ -1,0 +1,34 @@
+:host * {
+  box-sizing: border-box;
+}
+:host {
+  display: block;
+  position: relative;
+  cursor: pointer;
+}
+:host .hotspot {
+  border: 1px solid var(--border-color, #CCC);
+  border-radius: var(--border-radius, 5px);
+  padding: 5px 10px;
+}
+:host .flyout {
+  position: absolute;
+  border: 1px solid var(--border-color, #CCC);
+  min-width: 100%;
+  transform: translateY(-1px);
+  opacity: 0;
+  transition: opacity 150ms ease-out;
+  background-color: #fff;
+  height: 0;
+  overflow: hidden;
+}
+
+:host-context(.showOption) .hotspot {
+  border-bottom-left-radius: 0;
+  border-bottom-right-radius: 0;
+}
+:host-context(.showOption) .flyout {
+  opacity: 1;
+  z-index: 10000000000;
+  height: auto;
+}
