+++ src/components/form-controls/int-select/int-select.spec.ts
@@ -1,0 +1,7 @@
+import { IntSelect } from './int-select';
+
+describe('int-select', () => {
+  it('builds', () => {
+    expect(new IntSelect()).toBeTruthy();
+  });
+});
