--- src/components/form-controls/select/item/item.css
@@ -1,21 +1,0 @@
-:host {
-    display: flex;
-    font-family: inherit;
-    font-size: inherit;
-    cursor: pointer;
-    padding: 5px;
-}
-
-:host(:hover) {
-    background-color: rgb(226, 226, 226);
-}
-:host > * {
-    cursor: pointer;
-}
-:host input {
-    flex: 0 1 auto;
-    margin-right: 5px;
-}
-:host label {
-    flex: 1 auto;
-}
