--- src/components/form-controls/select/item/readme.md
@@ -1,24 +1,0 @@
-# int-select-item
-
-
-
-<!-- Auto Generated Below -->
-
-
-## Properties
-
-| Property | Attribute | Description | Type  | Default     |
-| -------- | --------- | ----------- | ----- | ----------- |
-| `value`  | `value`   |             | `any` | `undefined` |
-
-
-## Events
-
-| Event      | Description | Type               |
-| ---------- | ----------- | ------------------ |
-| `selected` |             | `CustomEvent<any>` |
-
-
-----------------------------------------------
-
-*Built with [StencilJS](https://stenciljs.com/)*
