--- src/components/form-controls/select/readme.md
@@ -1,26 +1,0 @@
-# int-select
-
-
-
-<!-- Auto Generated Below -->
-
-
-## Properties
-
-| Property      | Attribute     | Description | Type     | Default           |
-| ------------- | ------------- | ----------- | -------- | ----------------- |
-| `name`        | `name`        |             | `string` | `undefined`       |
-| `placeholder` | `placeholder` |             | `string` | `'Please select'` |
-| `value`       | `value`       |             | `any`    | `undefined`       |
-
-
-## Events
-
-| Event       | Description | Type               |
-| ----------- | ----------- | ------------------ |
-| `didChange` |             | `CustomEvent<any>` |
-
-
-----------------------------------------------
-
-*Built with [StencilJS](https://stenciljs.com/)*
